+++ react/components/SearchBar/components/SearchBar.js
@@ -50,7 +50,7 @@ const SearchBar = ({
   displayMode,
   minSearchTermLength,
   autocompleteFullWidth,
-  searchInput,
+  inputType,
 }) => {
   const intl = useIntl()
   const container = useRef()
@@ -207,7 +207,7 @@ const SearchBar = ({
                 submitOnIconClick={submitOnIconClick}
                 displayMode={displayMode}
                 openAutocompleteOnFocus={openAutocompleteOnFocus}
-                searchInput={searchInput}
+                inputType={inputType}
                 openMenu={openMenu}
                 inputErrorMessage={inputErrorMessage}
                 {...getInputProps({
@@ -310,8 +310,8 @@ SearchBar.propTypes = {
   minSearchTermLength: PropTypes.number,
   /** If true, the autocomplete will fill the whole window horizontally */
   autocompleteFullWidth: PropTypes.bool,
-  /** If true, the autocomplete will be an input type search */
-  searchInput: PropTypes.bool,
+  /** The type of the search input */
+  inputType: PropTypes.oneOf(['text', 'search']),
 }
 
 export default SearchBar