+++ CHANGELOG.md
@@ -7,9 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `searchInput` prop to `SearchBar`. If true, the autocomplete will be an input type search 
--  With searchInput=true it solves the problem when you click on the 'ok' button then disappears the results block on iOs platform
--  With searchInput=true it allows also to have a 'Search button' on the iOs virtual keyboard.
+- New prop `inputType` to `search-bar`.
 
 ## [3.131.3] - 2020-11-13
 ### Fixed